backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
shodan==1.30.1
dnspython<=2.0.0
beautifulsoup4==4.12.2
python-whois==0.8.0
ipwhois==1.2.0
gunicorn==21.2.0 